% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datas.R
\docType{data}
\name{xy}
\alias{xy}
\title{My data}
\format{
An object of class \code{data.frame} with 100 rows and 2 columns.
}
\usage{
xy
}
\arguments{
\item{y}{}
}
\value{

}
\description{
This is my first data
}
\keyword{datasets}
